MAPR-18234 MapRFileSystem error in hive.log

+++ ql/src/java/org/apache/hadoop/hive/ql/Context.java
@@ -503,7 +503,9 @@ public class Context {
     if (resDir != null) {
       try {
         FileSystem fs = resDir.getFileSystem(conf);
-        fs.delete(resDir, true);
+        if (fs.exists(resDir)) {
+          fs.delete(resDir, true);
+        }
       } catch (IOException e) {
         LOG.info("Context clear error: " + StringUtils.stringifyException(e));
       }
@@ -512,7 +514,9 @@ public class Context {
     if (resFile != null) {
       try {
         FileSystem fs = resFile.getFileSystem(conf);
-        fs.delete(resFile, false);
+        if (fs.exists(resFile)) {
+          fs.delete(resFile, false);
+        }
       } catch (IOException e) {
         LOG.info("Context clear error: " + StringUtils.stringifyException(e));
       }
